New communication
Changed the communication expected from the raspberry pi. This includes a check for which operating arduino the command is for, and a single number for the tray position. Also only 1 (open) or 0 (closed) for the servo. Also updated the response for a tray change to the new position in mm. 

Not tested on arduino so might contain (a lot) of syntax errors.